Moved tests to more sensible files
