+++ Tests/Engine.test/JourneyPipeline.Test.cs
@@ -33,118 +33,6 @@ public class JourneyPipelineTests
         Assert.Throws<InvalidOperationException>(() => new JourneyPipeline(grid, cities, new StubRouter([])));
     }
 
-    [Fact]
-    public void BoundingBox_ReturnsExpectedMinMax_NormalCell()
-    {
-        var grid = new SpawnGrid([], new Position(0.0, 0.0), latSize: 2.0, lonSize: 4.0);
-        var cell = new GridCell(spawnable: true, new Position(10.0, 20.0));
-
-        var (min, max) = grid.GetBoundingBox(cell);
-
-        Assert.Equal(8.0, min.Longitude);
-        Assert.Equal(19.0, min.Latitude);
-        Assert.Equal(12.0, max.Longitude);
-        Assert.Equal(21.0, max.Latitude);
-    }
-
-    [Fact]
-    public void BoundingBox_ReturnsExpectedMinMax_UnitCell()
-    {
-        var grid = new SpawnGrid([], new Position(0.0, 0.0), latSize: 1.0, lonSize: 1.0);
-        var cell = new GridCell(spawnable: true, new Position(0.0, 0.0));
-
-        var (min, max) = grid.GetBoundingBox(cell);
-
-        Assert.Equal(-0.5, min.Longitude);
-        Assert.Equal(-0.5, min.Latitude);
-        Assert.Equal(0.5, max.Longitude);
-        Assert.Equal(0.5, max.Latitude);
-    }
-
-    [Fact]
-    public void BoundingBox_ContainsCenterPoint()
-    {
-        var grid = new SpawnGrid([], new Position(0.0, 0.0), latSize: 2.0, lonSize: 2.0);
-        var cell = new GridCell(spawnable: true, new Position(5.0, 5.0));
-
-        var (min, max) = grid.GetBoundingBox(cell);
-
-        Assert.InRange(cell.Centerpoint.Longitude, min.Longitude, max.Longitude);
-        Assert.InRange(cell.Centerpoint.Latitude, min.Latitude, max.Latitude);
-    }
-
-    [Fact]
-    public void GenerateGrid_CellInsideWetPolygon_NotSpawnable()
-    {
-        var polygon = new List<Position>
-        {
-            new(-1, -1), new(1, -1), new(1, 1), new(-1, 1),
-        };
-        var wetPolygon = new List<Position>
-        {
-            new(-0.1, -0.1), new(0.1, -0.1), new(0.1, 0.1), new(-0.1, 0.1),
-        };
-        var stationPolygons = new List<Position>
-        {
-            new(0.8, 0.8),
-        };
-
-        var grid = Polygooner.GenerateGrid(0.05, [polygon], [wetPolygon], stationPolygons);
-        var center = grid.Cells.SelectMany(r => r)
-            .FirstOrDefault(c => Math.Abs(c.Centerpoint.Longitude) < 0.05
-                              && Math.Abs(c.Centerpoint.Latitude) < 0.05);
-
-        Assert.NotNull(center);
-        Assert.False(center.Spawnable);
-    }
-
-    [Fact]
-    public void GenerateGrid_CellOutsideWetPolygon_Spawnable()
-    {
-        var polygon = new List<Position>
-        {
-            new(-1, -1), new(1, -1), new(1, 1), new(-1, 1),
-        };
-        var wetPolygon = new List<Position>
-        {
-            new(-0.1, -0.1), new(0.1, -0.1), new(0.1, 0.1), new(-0.1, 0.1),
-        };
-        var stationPolygons = new List<Position>
-        {
-            new(0.8, 0.8),
-        };
-
-        var grid = Polygooner.GenerateGrid(0.05, [polygon], [wetPolygon], stationPolygons);
-        var farCell = grid.Cells.SelectMany(r => r)
-            .FirstOrDefault(c => c.Centerpoint.Longitude > 0.5
-                              && c.Centerpoint.Latitude > 0.5);
-
-        Assert.NotNull(farCell);
-        Assert.True(farCell.Spawnable);
-    }
-
-    [Fact]
-    public void GenerateGrid_EntireAreaWet_NoCellsSpawnable()
-    {
-        var polygon = new List<Position>
-        {
-            new(-1, -1), new(1, -1), new(1, 1), new(-1, 1),
-        };
-        var wetPolygon = new List<Position>
-        {
-            new(-2, -2), new(2, -2), new(2, 2), new(-2, 2),
-        };
-        var stationPolygons = new List<Position>
-        {
-            new(0.8, 0.8),
-        };
-
-        var grid = Polygooner.GenerateGrid(0.1, [polygon], [wetPolygon], stationPolygons);
-        var anySpawnable = grid.Cells.SelectMany(r => r).Any(c => c.Spawnable);
-
-        Assert.False(anySpawnable);
-    }
-
     private class StubRouter(float[] distances) : IMatrixRouter
     {
         private readonly float[] _distances = distances;
